Create Result Data.py
Use this file for extracting data from students result marksheet from the bangladesh education system board website, input the compulsory fields in the excel file like - Roll, Reg, Board, Year and save the data and close the file and run the python file in the same folder by double clicking, and remember to change the excel file path from the python file to the excel file path from your file location. After running the code will automatically go to the board website and fill the data and complete the captcha and  scrape the data from the website to the excel file.